+++ build_config.glibc-2.14.rb
@@ -37,13 +37,13 @@ end
 unless ARGV.include? 'test:bintest'
 
   MRuby::Build.new('x86_64-pc-linux-gnu') do |conf|
-    toolchain :gcc
+    toolchain :clang
 
     gem_config(conf)
   end
 
   MRuby::CrossBuild.new('i686-pc-linux-gnu') do |conf|
-    toolchain :gcc
+    toolchain :clang
 
     [conf.cc, conf.cxx, conf.linker].each do |cc|
       cc.flags << '-m32'