+++ build_config.musl.rb
@@ -25,7 +25,7 @@ def gem_config(conf)
 end
 
 MRuby::Build.new do |conf|
-  toolchain :gcc
+  toolchain :clang
 
   conf.enable_bintest
   conf.enable_debug
@@ -37,7 +37,7 @@ end
 unless ARGV.include? 'test:bintest'
 
   MRuby::Build.new('x86_64-pc-linux-busybox') do |conf|
-    toolchain :gcc
+    toolchain :clang
 
     gem_config(conf)
   end