+++ build_config.glibc-2.14.rb
@@ -50,32 +50,32 @@ MRuby::CrossBuild.new('i686-pc-linux-gnu') do |conf|
   gem_config(conf)
 end
 
-MRuby::CrossBuild.new('x86_64-apple-darwin14') do |conf|
+MRuby::CrossBuild.new('x86_64-apple-darwin15') do |conf|
   toolchain :clang
 
   [conf.cc, conf.linker].each do |cc|
-    cc.command = 'x86_64-apple-darwin14-clang'
+    cc.command = 'x86_64-apple-darwin15-clang'
   end
-  conf.cxx.command      = 'x86_64-apple-darwin14-clang++'
-  conf.archiver.command = 'x86_64-apple-darwin14-ar'
+  conf.cxx.command      = 'x86_64-apple-darwin15-clang++'
+  conf.archiver.command = 'x86_64-apple-darwin15-ar'
 
   conf.build_target     = 'x86_64-pc-linux-gnu'
-  conf.host_target      = 'x86_64-apple-darwin14'
+  conf.host_target      = 'x86_64-apple-darwin15'
 
   gem_config(conf)
 end
 
-MRuby::CrossBuild.new('i386-apple-darwin14') do |conf|
+MRuby::CrossBuild.new('i386-apple-darwin15') do |conf|
   toolchain :clang
 
   [conf.cc, conf.linker].each do |cc|
-    cc.command = 'i386-apple-darwin14-clang'
+    cc.command = 'i386-apple-darwin15-clang'
   end
-  conf.cxx.command      = 'i386-apple-darwin14-clang++'
-  conf.archiver.command = 'i386-apple-darwin14-ar'
+  conf.cxx.command      = 'i386-apple-darwin15-clang++'
+  conf.archiver.command = 'i386-apple-darwin15-ar'
 
   conf.build_target     = 'i386-pc-linux-gnu'
-  conf.host_target      = 'i386-apple-darwin14'
+  conf.host_target      = 'i386-apple-darwin15'
 
   gem_config(conf)
 end