+++ CHANGELOG.md
@@ -14,6 +14,8 @@
    $ ff -s type=db
    ```
 
+5. Performance enhancements
+
 ### 1.4.1 (15.03.2017)
 
 1. Colorized error output.