+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 ## Release Notes: _fifa_
 
-### 1.4.2 (not yet released)
+### 1.4.2 (31.05.2017)
 
 1. Bug fixes for edge cases like if type attribute is missing.